Create apply coupon

+++ src/main/java/id/holigo/services/holigocouponservice/domain/Coupon.java
@@ -52,7 +52,7 @@ public class Coupon {
 
     private Boolean isPublic;
 
-    private Short quantity;
+    private Integer quantity;
 
     private Timestamp validAt;
